Add input for Seed, impl Eq for Seed, impl ToPrivKey for Seed
Originally tried to return an array, but slices aren't returnable due
tonot living past the life of the function. Luckily blake2 has a
`finalized_box` implementation, which allocates to the heap, resulting
in a returnable data structure.

Tried to derive Eq initially but it didn't work. Then impl'd PartialEq
and Eq directly. Will try derive(PartialEq, Eq) next time. Got some
practice on implementing it anyway.

The hex crate is pretty handy, but still didn't allow me to implement
LowerHex and UpperHex on Seed struct. Was able to avoid this by adding a
`?` to the end of the formatting, ex. `{:X?}`